% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{find_all_gen}
\alias{find_all_gen}
\title{Return a vector of positions defining new generations in a pedigree}
\usage{
find_all_gen(ped, start = 1)
}
\arguments{
\item{ped}{A data.frame made up of four
variables, id, Female, Male, Observed.}

\item{start}{The row to start from}
}
\description{
Takes a data.frame made up of four variables,
id, Female, Male, Observed, and returns a
vector of rows where the next generation starts.
}

