% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{find_gen}
\alias{find_gen}
\title{Finds rows in a data.frame pedigree where the
generation changes}
\usage{
find_gen(ped, start = 1)
}
\arguments{
\item{ped}{A data.frame made up of four
variables, id, Female, Male, Observed.}

\item{start}{The row to start from

Some useful keyboard shortcuts for package authoring:

  Build and Reload Package:  'Ctrl + Shift + B'
  Check Package:             'Ctrl + Shift + E'
  Test Package:              'Ctrl + Shift + T'}
}
\description{
Takes a data.frame made up of four variables,
id, Female, Male, Observed, and returns the
row where the generation changes.
}

