% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{make_gen}
\alias{make_gen}
\title{Return a vector describing each generation of an entry in a pedigree}
\usage{
make_gen(ped)
}
\arguments{
\item{ped}{A data.frame made up of four
variables, id, Female, Male, Observed.}
}
\description{
Takes a data.frame pedigree and returns a vector of the same length defining which generation each row of the pedigree fits.
}

