% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_plot.R
\name{make_ped_plot}
\alias{make_ped_plot}
\title{Make a pedigree plot}
\usage{
make_ped_plot(ped)
}
\arguments{
\item{ped}{A data.frame made up of four
variables, id, Female, Male, Observed.}
}
\description{
Takes a pedigree data.frame and returns a pedigree plot
}

